rename consts file to movie-variables
